Adjusted and fixes some bugs for calendar page

+++ database/migrations/008_create_events.sql
@@ -1,5 +1,6 @@
 CREATE TABLE IF NOT EXISTS events (
-    id INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+
     title VARCHAR(180) NOT NULL,
     description TEXT NOT NULL,
     event_date DATE NOT NULL,
@@ -9,15 +10,31 @@ CREATE TABLE IF NOT EXISTS events (
     organizer VARCHAR(160) NOT NULL,
     website_url VARCHAR(500) NULL,
     material_request TEXT NULL,
-    status ENUM('pending','approved','rejected') NOT NULL DEFAULT 'pending',
-    submitted_by INT UNSIGNED NOT NULL,
-    reviewed_by INT UNSIGNED NULL,
+
+    status ENUM('pending', 'approved', 'rejected')
+        NOT NULL DEFAULT 'pending',
+
+    submitted_by BIGINT UNSIGNED NOT NULL,
+    reviewed_by BIGINT UNSIGNED NULL,
     review_reason VARCHAR(500) NULL,
     reviewed_at DATETIME NULL,
+
     created_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    updated_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
+    updated_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP
+        ON UPDATE CURRENT_TIMESTAMP,
+
     INDEX idx_events_status_date (status, event_date),
     INDEX idx_events_submitter_status (submitted_by, status),
-    CONSTRAINT fk_events_submitter FOREIGN KEY (submitted_by) REFERENCES users(id) ON DELETE CASCADE,
-    CONSTRAINT fk_events_reviewer FOREIGN KEY (reviewed_by) REFERENCES users(id) ON DELETE SET NULL
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
+
+    CONSTRAINT fk_events_submitter
+        FOREIGN KEY (submitted_by)
+        REFERENCES users(id)
+        ON DELETE CASCADE,
+
+    CONSTRAINT fk_events_reviewer
+        FOREIGN KEY (reviewed_by)
+        REFERENCES users(id)
+        ON DELETE SET NULL
+) ENGINE=InnoDB
+DEFAULT CHARSET=utf8mb4
+COLLATE=utf8mb4_unicode_ci;
